requirements.txt
pytest-playwright==0.7.1
datetime
babel